* m4/libtool.m4 (AC_PROG_SED): Don't break test loop early.

+++ m4/libtool.m4
@@ -5970,7 +5970,7 @@ lt_ac_count=0
 # Add /usr/xpg4/bin/sed as it is typically found on Solaris
 # along with /bin/sed that truncates output.
 for lt_ac_sed in $lt_ac_sed_list /usr/xpg4/bin/sed; do
-  test ! -f $lt_ac_sed && break
+  test ! -f $lt_ac_sed && continue
   cat /dev/null > conftest.in
   lt_ac_count=0
   $ECHO $ECHO_N "0123456789$ECHO_C" >conftest.in
